Add files via upload
This is only contains the sheets and content for the Module 1 challenge - Outcome Based on Goal and Outcomes Based on Launch Date. If you are looking for the other sheets and charts, please refer to the excel document KickStarter Data_Louise Hsu . I did not include the extra sheets in this document because I thought it would be easier to navigate.